GPA writes windows events to 'Windows Azure' instead of 'Application'.
